# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  handlers.LoginResponse:
    properties:
      access_token:
        type: string
      user_id:
        type: string
    type: object
...
  services.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
    type: object
  services.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
      - email
      - password
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
  title: Todo app API
  version: '1.0'
paths:
  /login:
    post:
      consumes:
        - application/json
      parameters:
        - description: user credential
          in: body
          name: credential
          required: true
          schema:
            $ref: '#/definitions/services.LoginRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        '400':
          description: '{"error": "Invalid request"}'
          schema:
            $ref: '#/definitions/gin.H'
        '401':
          description: '{"error": "Invalid email or password"}'
          schema:
            $ref: '#/definitions/gin.H'
        '500':
          description: '{"error": "Internal server error"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: Login a user
      tags:
        - Auth
  /logout:
    post:
      produces:
        - application/json
      responses:
        '200':
          description: '{"message": "Logged out"}'
          schema:
            $ref: '#/definitions/gin.H'
        '500':
          description: '{"error": "Internal server error"}'
          schema:
            $ref: '#/definitions/gin.H'
      security:
        - BearerAuth: []
      summary: Logout a user
      tags:
        - Auth
...
  /register:
    post:
      consumes:
        - application/json
      parameters:
        - description: user credential
          in: body
          name: credential
          required: true
          schema:
            $ref: '#/definitions/services.RegisterRequest'
      produces:
        - application/json
      responses:
        '200':
          description: '{"message": "User registered"}'
          schema:
            $ref: '#/definitions/gin.H'
        '400':
          description: '{"error": "Invalid request"}'
          schema:
            $ref: '#/definitions/gin.H'
        '409':
          description: '{"error": "User already registered"}'
          schema:
            $ref: '#/definitions/gin.H'
        '500':
          description: '{"error": "The server encountered unexpected error"}'
          schema:
            $ref: '#/definitions/gin.H'
      summary: Register an user
      tags:
        - Auth
...
swagger: '2.0'